initial restructure of server files
